Get rid of mutators file
